+++ templates/element/info/release-notes.php
@@ -2,7 +2,7 @@
 <?php include('version.php'); ?>
 <p></p>
 <h3>Release Notes</h3>
-If you are curious about the latest technical developments and enhancements, check our release notes on Github:
-<a href="https://github.com/acdh-oeaw/dhcr-main/blob/prod/RELEASE_NOTES.md" target="_blank">
-    https://github.com/acdh-oeaw/dhcr-main/blob/prod/RELEASE_NOTES.md
+If you are curious about the latest technical developments and enhancements, check our release notes on Github:<br>
+<a href="https://github.com/acdh-oeaw/dhcr-main/blob/prod/documentation/RELEASE_NOTES.md" target="_blank">
+    https://github.com/acdh-oeaw/dhcr-main/blob/prod/documentation/RELEASE_NOTES.md
 </a>