+++ documentation/RELEASE_NOTES.md
@@ -5,7 +5,7 @@
 
 ### New features
 - Reimplemented filter options for Tadirah Technique or Object #20
-- Created technical documentation #33
+- Started with technical documentation #33
 
 ### Changes
 - Execute filter by GO-button #136